// frontend_user/types/supabase.ts
export type Database = {
    public: {
        Tables: {
// ...
            feedback: {
                Row: {
                    chat_id: string | null
                    comment: string | null
                    created_at: string
                    id: number
                    like: boolean | null
                    message_id: string | null
                    user_id: string | null
                }
                Insert: {
                    chat_id?: string | null
                    comment?: string | null
                    created_at?: string
                    id?: number
                    like?: boolean | null
                    message_id?: string | null
                    user_id?: string | null
                }
                Update: {
                    chat_id?: string | null
                    comment?: string | null
                    created_at?: string
                    id?: number
                    like?: boolean | null
                    message_id?: string | null
                    user_id?: string | null
                }
                Relationships: [
                    {
                        foreignKeyName: "public_feedback_chat_id_fkey"
                        columns: ["chat_id"]
                        isOneToOne: false
                        referencedRelation: "chat_session"
                        referencedColumns: ["chat_id"]
                    },
                    {
                        foreignKeyName: "public_feedback_message_id_fkey"
                        columns: ["message_id"]
                        isOneToOne: false
                        referencedRelation: "chat_message"
                        referencedColumns: ["message_id"]
                    },
                    {
                        foreignKeyName: "public_feedback_user_id_fkey"
                        columns: ["user_id"]
                        isOneToOne: false
                        referencedRelation: "profiles"
                        referencedColumns: ["id"]
                    },
                ]
            }
// ...
        }
// ...
    }
}